backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pymongo==4.10.1
python-dotenv==1.1.1
python-multipart==0.0.7
pydantic==2.11.7
numpy==2.0.0
scikit-learn==1.5.1
emergentintegrations==0.1.0
pyyaml==6.0.2
loguru==0.7.2